requirements.txt
pymongo
pyodbc
colorama
openai
anthropic
groq
google-genai
tiktoken
django==4.2.7
djangorestframework==3.14.0